메인액티비티 변경, 학생목록 액티비티 생성
