+++ app/src/main/java/kr/co/tjeit/lecturemanager/LoginActivity.java
@@ -4,7 +4,6 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
-import android.widget.Toast;
 
 import com.facebook.CallbackManager;
 import com.facebook.FacebookCallback;
@@ -61,7 +60,7 @@ public class LoginActivity extends BaseActivity {
         loginBtn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent = new Intent(LoginActivity.this, MainActivity.class);
+                Intent intent = new Intent(LoginActivity.this, StudentListActivity.class);
                 startActivity(intent);
                 finish();
             }
@@ -99,7 +98,7 @@ public class LoginActivity extends BaseActivity {
 
                     ContextUtil.login(mContext, tempUser);
 
-                    Intent intent = new Intent(mContext, MainActivity.class);
+                    Intent intent = new Intent(mContext, StudentListActivity.class);
                     startActivity(intent);
 
                 }
@@ -166,7 +165,7 @@ public class LoginActivity extends BaseActivity {
 
                     ContextUtil.login(mContext, tempUser);
 
-                    Intent intent = new Intent(mContext, MainActivity.class);
+                    Intent intent = new Intent(mContext, StudentListActivity.class);
                     startActivity(intent);
                 }
             });
