+++ app/src/main/java/kr/co/tjeit/lecturemanager/MainActivity.java
@@ -1,25 +1,9 @@
 package kr.co.tjeit.lecturemanager;
 
-import android.content.DialogInterface;
-import android.content.Intent;
+import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.app.AlertDialog;
-import android.view.View;
-import android.widget.AdapterView;
-import android.widget.Button;
-import android.widget.ListView;
-
-import kr.co.tjeit.lecturemanager.adapter.StudentAdapter;
-import kr.co.tjeit.lecturemanager.util.GlobalData;
-
-// 저만의 브런치 입니다. (조경진)
 
 public class MainActivity extends BaseActivity {
-
-    private ListView studentListView;
-    private android.widget.Button myProfileBtn;
-
-    StudentAdapter mAdapter;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -28,72 +12,20 @@ public class MainActivity extends BaseActivity {
         bindViews();
         setupEvents();
         setValues();
-
-
-
     }
 
     @Override
     public void setupEvents() {
 
-        myProfileBtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                Intent intent = new Intent(mContext, MyProfileActivity.class);
-                startActivity(intent);
-            }
-        });
-
-        studentListView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-//                Toast.makeText(MainActivity.this, position+"번 줄", Toast.LENGTH_SHORT).show();
-
-                Intent myIntent = new Intent(MainActivity.this, ViewStudentInfoActivity.class);
-                myIntent.putExtra("사용자정보", GlobalData.allUsers.get(position));
-                startActivity(myIntent);
-
-            }
-        });
-
-        studentListView.setOnItemLongClickListener(new AdapterView.OnItemLongClickListener() {
-            @Override
-            public boolean onItemLongClick(AdapterView<?> parent, View view, final int position, long id) {
-
-//                Toast.makeText(MainActivity.this, position + "번 학생 삭제 요청", Toast.LENGTH_SHORT).show();
-
-                AlertDialog.Builder myBuilder = new AlertDialog.Builder(MainActivity.this);
-                myBuilder.setTitle("삭제 확인");
-                myBuilder.setMessage("정말 삭제하시겠습니까?");
-                myBuilder.setPositiveButton("확인", new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
-
-
-                    }
-                });
-                myBuilder.setNegativeButton("취소", null);
-                myBuilder.show();
-
-                return false;
-            }
-        });
-
-
     }
 
     @Override
     public void setValues() {
-        mAdapter = new StudentAdapter(mContext, GlobalData.allUsers);
-        studentListView.setAdapter(mAdapter);
+
     }
 
     @Override
     public void bindViews() {
 
-        this.myProfileBtn = (Button) findViewById(R.id.myProfileBtn);
-        studentListView = (ListView) findViewById(R.id.studentListView);
     }
 }
-
-
